Spanish layout
So I'm going to be clear, I do not know much about coding but I managed to edit this code and made it work for a Spanish layout just by changing some things around line 203 the rest is the same. I only changed what I think are essential characters like ()[]{} (some of them worked without changing anything) so for example what I think are useless characters like ñ or ¡¿ are not included (I tried but could not do it I do not know why).